refactor(communities): Replace hash-based navigation with proper SvelteKit routing
- Replace manual hash listeners with route parameters /communities/[section]
- Create new route structure for continent and organization sections
- Update communities page to redirect to new structure with backward compatibility
- Add server-side validation for continent sections
- Remove client-side hash manipulation in favor of proper navigation
- Update e2e tests to match new routing patterns
- Support dynamic organization sections with client-side validation
- Maintain chart functionality and section filtering with new routing
